pyproject.toml
[project]
name = "ai-finance-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "langchain>=1.2.3",
    "langchain-community>=0.3.20",
    "pypdf>=5.1.0",
    "docx2txt>=0.8",
    "langchain-classic>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.0.0",
    "pinecone>=5.0.0",
    "langgraph>=0.2.0",
    "yfinance>=1.0",
    "python-dateutil>=2.8.2",
    "gradio>=6.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
